# specs/001-festival-timeline-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Festival Timeline Management API
  description: REST API for collaborative festival event planning with real-time timeline visualization
  version: 1.0.0
  contact:
    name: Festival Timeline Team

servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.festival-timeline.app/api
    description: Production server

tags:
  - name: Auth
    description: Authentication and user registration
  - name: Events
    description: Event management operations
  - name: Categories
    description: Category management operations
  - name: Users
    description: User management operations
  - name: Export
    description: Data export operations

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@festival.app"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [Admin, Organizer, Viewer]
          example: "Organizer"
        oauthProvider:
          type: string
          enum: [google, microsoft]
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt

    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Venue Setup"
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: "#4A90E2"
        createdBy:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdBy
        - createdAt

    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: "Setup main stage"
        date:
          type: string
          format: date
          example: "2025-12-25"
        description:
          type: string
          nullable: true
          maxLength: 10000
          example: "Install lighting and sound equipment on main stage"
        categoryId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        assignedPerson:
          type: string
          nullable: true
          maxLength: 255
          example: "Alice Johnson"
        status:
          type: string
          enum: [Not Started, In Progress, Completed, Blocked]
          default: "Not Started"
          example: "In Progress"
        priority:
          type: string
          enum: [High, Medium, Low]
          default: "Medium"
          example: "High"
        createdBy:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - date
        - categoryId
        - status
        - priority
        - createdBy
        - createdAt
        - updatedAt

    EventWithDetails:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            category:
              $ref: '#/components/schemas/Category'
            createdByUser:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string

    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "Event title is required"
        code:
          type: string
          example: "VALIDATION_ERROR"
      required:
        - error
        - message

paths:
  # Authentication Endpoints
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Create a new user account with email and password (FR-024 - open self-registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
              required:
                - email
                - password
                - name
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    description: JWT access token
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/google:
    get:
      tags: [Auth]
      summary: Initiate Google OAuth flow
      description: Redirects to Google OAuth consent screen
      responses:
        '302':
          description: Redirect to Google OAuth

  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      description: Handles Google OAuth callback and creates/updates user
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
      responses:
        '302':
          description: Redirect to frontend with JWT token

  /auth/microsoft:
    get:
      tags: [Auth]
      summary: Initiate Microsoft OAuth flow
      description: Redirects to Microsoft OAuth consent screen
      responses:
        '302':
          description: Redirect to Microsoft OAuth

  /auth/microsoft/callback:
    get:
      tags: [Auth]
      summary: Microsoft OAuth callback
      description: Handles Microsoft OAuth callback and creates/updates user
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
      responses:
        '302':
          description: Redirect to frontend with JWT token

  # Event Endpoints
  /events:
    get:
      tags: [Events]
      summary: Get all events
      description: Retrieve all events with optional filtering (FR-009)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Not Started, In Progress, Completed, Blocked]
          description: Filter by status
        - in: query
          name: priority
          schema:
            type: string
            enum: [High, Medium, Low]
          description: Filter by priority
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
          description: Filter by category
        - in: query
          name: assignedPerson
          schema:
            type: string
          description: Filter by assigned person
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Filter events from this date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Filter events until this date
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [date, urgency, priority]
            default: date
          description: Sort order (FR-008)
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventWithDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Events]
      summary: Create a new event
      description: Create a new event (FR-001, FR-002)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                date:
                  type: string
                  format: date
                description:
                  type: string
                  maxLength: 10000
                categoryId:
                  type: string
                  format: uuid
                assignedPerson:
                  type: string
                  maxLength: 255
                status:
                  type: string
                  enum: [Not Started, In Progress, Completed, Blocked]
                  default: "Not Started"
                priority:
                  type: string
                  enum: [High, Medium, Low]
                  default: "Medium"
              required:
                - title
                - date
                - categoryId
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithDetails'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /events/{id}:
    get:
      tags: [Events]
      summary: Get event by ID
      description: Retrieve detailed information about a specific event (FR-004)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithDetails'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Events]
      summary: Update an event
      description: Update event details (FR-005)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                date:
                  type: string
                  format: date
                description:
                  type: string
                  maxLength: 10000
                categoryId:
                  type: string
                  format: uuid
                assignedPerson:
                  type: string
                  maxLength: 255
                status:
                  type: string
                  enum: [Not Started, In Progress, Completed, Blocked]
                priority:
                  type: string
                  enum: [High, Medium, Low]
                updatedAt:
                  type: string
                  format: date-time
                  description: For optimistic locking (FR-018)
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithDetails'
        '409':
          description: Conflict - concurrent edit detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Event not found

    delete:
      tags: [Events]
      summary: Delete an event
      description: Delete an event (FR-006 - requires confirmation at UI level)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
        '401':
          description: Unauthorized

  # Category Endpoints
  /categories:
    get:
      tags: [Categories]
      summary: Get all categories
      description: Retrieve all event categories
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags: [Categories]
      summary: Create a new category
      description: Create a new category (FR-021 - any authenticated user can create)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
              required:
                - name
                - color
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '409':
          description: Category name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories/{id}:
    put:
      tags: [Categories]
      summary: Update a category
      description: Update category (FR-022 - only creator can modify)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '403':
          description: Forbidden - not creator
        '404':
          description: Category not found

    delete:
      tags: [Categories]
      summary: Delete a category
      description: Delete category (FR-022 - only creator can delete, cannot delete if events exist)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Category deleted
        '403':
          description: Forbidden - not creator
        '409':
          description: Conflict - category has events
        '404':
          description: Category not found

  # Export Endpoints
  /export/timeline-pdf:
    post:
      tags: [Export]
      summary: Export timeline as PDF
      description: Generate PDF of timeline visualization (FR-013, SC-007 - <5s)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                categories:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Optional category filter
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '500':
          description: Export failed

  /export/timeline-image:
    post:
      tags: [Export]
      summary: Export timeline as image
      description: Generate PNG image of timeline (FR-014, SC-007 - <5s)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                categories:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary

  /export/events-csv:
    get:
      tags: [Export]
      summary: Export events as CSV
      description: Export all events to CSV format (FR-015, SC-008 - 100% accuracy)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                example: |
                  id,title,date,description,category,assignedPerson,status,priority,createdBy,createdAt,updatedAt
                  770e8400...,Setup main stage,2025-12-25,Install equipment,Venue Setup,Alice,In Progress,High,550e8400...,2025-10-01,2025-10-18

  /export/events-excel:
    get:
      tags: [Export]
      summary: Export events as Excel
      description: Export all events to Excel format (FR-015, SC-008 - 100% accuracy)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  # User Management (Admin only)
  /users:
    get:
      tags: [Users]
      summary: Get all users
      description: List all users (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Admin only

  /users/{id}/role:
    patch:
      tags: [Users]
      summary: Update user role
      description: Change user role (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [Admin, Organizer, Viewer]
              required:
                - role
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden - Admin only

security:
  - BearerAuth: []